% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_noerror.R
\name{boot_unadj_point_exact.f}
\alias{boot_unadj_point_exact.f}
\title{Bootstrap confidence interval of the mediation effect based on exact expressions (no measurement error)}
\usage{
boot_unadj_point_exact.f(
  Outcome.formula,
  Mediator.formula,
  data,
  Tname,
  Dname,
  Aname,
  Mname,
  Wname,
  a0,
  a1,
  cw,
  t.list = c(10, 20, 30, 40),
  R = 50
)
}
\description{
Bootstrap confidence interval of the mediation effect based on exact expressions (no measurement error)
}
