% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exact.R
\name{bootci.f}
\alias{bootci.f}
\title{Bootstrap confidence interval of the mediation effect based on exact expressions by all proposed estimators}
\usage{
bootci.f(
  Outcome.formula,
  Outcome.formula.rrc,
  Mediator.formula,
  ME1.formula,
  ME2.formula,
  data,
  Time_sep,
  Iname,
  Tname,
  Dname,
  Aname,
  Asname,
  Mname,
  Wname,
  a0,
  a1,
  cw,
  t.list = c(10, 20, 30, 40),
  R = 50
)
}
\description{
Bootstrap confidence interval of the mediation effect based on exact expressions by all proposed estimators
}
