% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exact.R
\name{combined.f}
\alias{combined.f}
\title{Point estimation of the mediation effect based on exact expressions by all proposed estimators}
\usage{
combined.f(
  Outcome.formula,
  Outcome.formula.rrc,
  Mediator.formula,
  ME1.formula,
  ME2.formula,
  data,
  Time_sep,
  Iname,
  Tname,
  Dname,
  Aname,
  Asname,
  Mname,
  Wname,
  a0,
  a1,
  cw,
  t.list
)
}
\description{
Point estimation of the mediation effect based on exact expressions by all proposed estimators
}
