% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cox.R
\name{cox_orc.f}
\alias{cox_orc.f}
\title{ORC method to obtain the Cox model coefficients}
\usage{
cox_orc.f(data, Iname, Tname, Dname, Aname, Asname, Wname)
}
\arguments{
\item{data}{dataset.}

\item{Iname}{name of the indicator on the main/validation study (I=1 for validation study sample, I=0 for main study samples).}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}
}
\value{
ORC estimates of the Cox regression coefficients.
}
\description{
ORC method to obtain the Cox model coefficients
}
