% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cox.R
\name{cox_rrc.f}
\alias{cox_rrc.f}
\title{RRC method to obtain Cox model coefficients}
\usage{
cox_rrc.f(data, Time_sep, Iname, Tname, Dname, Aname, Asname, Wname)
}
\arguments{
\item{data}{dataset.}

\item{Iname}{name of the indicator on the main/validation study (I=1 for validation study sample, I=0 for main study samples).}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}
}
\value{
RRC estimates of the Cox regression coefficients.
}
\description{
RRC method to obtain Cox model coefficients
}
