% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_approx.R
\name{get_estimated_effect}
\alias{get_estimated_effect}
\title{Obtain the estimated mediation effect measures based on the approximate expressions under a rare
outcome assumption}
\usage{
get_estimated_effect(res, Mname, Aname, Wname, a0 = 0, a1 = 1, cw = 0)
}
\arguments{
\item{res}{results from the unadj.f, orc1.f, orc2.f, or rrc.f}

\item{Mname}{name of the mediator}

\item{Aname}{name of the true exposure}

\item{Wname}{names of the covariates}

\item{a0}{baseline exposure level in the mediation effect measures}

\item{a1}{active exposure level in the mediation effect measures}

\item{cw}{levels of the covariates}
}
\value{
estimation of mediation effect measures
}
\description{
Obtain the estimated mediation effect measures based on the approximate expressions under a rare
outcome assumption
}
