% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_approx.R
\name{mediate_approx}
\alias{mediate_approx}
\title{Mediation analysis based on the approximate expressions under a rare outcome assumption.}
\usage{
mediate_approx(
  data,
  Iname,
  Tname,
  Dname,
  Aname,
  Asname,
  Mname,
  Wname,
  Time_sep,
  a0,
  a1,
  cw
)
}
\arguments{
\item{data}{dataset.}

\item{Iname}{name of the indicator on the main/validation study (I=1 for validation study sample, I=0 for main study samples).}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}

\item{Aname}{name of the true exposure}

\item{Asname}{name of the mismeasured exposure}

\item{Mname}{name of the mediator}

\item{Wname}{names of the covariates}

\item{Time_sep}{a vector of time split points for RRC method.}

\item{a0}{baseline exposure level in the mediation effect measures}

\item{a1}{active exposure level in the mediation effect measures}

\item{cw}{levels of the covariates}
}
\value{
estimation of mediation effect measures based on the unadjusted method, ORC1, ORC2, and RRC
}
\description{
Mediation analysis based on the approximate expressions under a rare outcome assumption.
}
\examples{
## attach example dataset
attach(example_data)

## specify the exposure, mediator, outcome, and covariates
Iname="I" # name of the indicator of main/validation study samples
Tname="Time" # name of observed failure time
Dname="Event" # name of failure indicator
Aname="A" # name of true exposure
Asname = "As" # name of mismeasured exposure
Mname = "M" # name of mediator
Wname="W" # name of covariates

## specify split points for RRC method
Time_sep=c(0,15,25,35)

## specify the conditional values in the mediation effect measures
a0=0; a1=1; cw=0

## run the following code to perform the mediation analysis with approximate expressions
## mediate_approx(example_data,Iname,Tname,Dname,Aname,
##                Asname,Mname,Wname,Time_sep,a0,a1,cw)
}
