% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exact.R
\name{mediate_exact}
\alias{mediate_exact}
\title{Mediation analysis based on the exact expressions}
\usage{
mediate_exact(
  data,
  Iname,
  Tname,
  Dname,
  Aname,
  Asname,
  Mname,
  Wname,
  t.list,
  Time_sep,
  a0,
  a1,
  cw,
  R = 50
)
}
\arguments{
\item{data}{dataset.}

\item{Iname}{name of the indicator on the main/validation study (I=1 for validation study sample, I=0 for main study samples).}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}

\item{Aname}{name of the true exposure}

\item{Asname}{name of the mismeasured exposure}

\item{Mname}{name of the mediator}

\item{Wname}{names of the covariates}

\item{t.list}{a list of time points for the mediation effect measures}

\item{Time_sep}{a vector of time split points for RRC method.}

\item{a0}{baseline exposure level in the mediation effect measures}

\item{a1}{active exposure level in the mediation effect measures}

\item{cw}{levels of the covariates}

\item{R}{number of bootstrap numbers (default 50)}
}
\value{
estimation of mediation effect measures based on the unadjusted method, ORC1, ORC2, and RRC
}
\description{
Mediation analysis based on the exact expressions
}
\examples{
## attach example dataset
attach(example_data)

## specify the exposure, mediator, outcome, and covariates
Iname="I" # name of the indicator of main/validation study samples
Tname="Time" # name of observed failure time
Dname="Event" # name of failure indicator
Aname="A" # name of true exposure
Asname = "As" # name of mismeasured exposure
Mname = "M" # name of mediator
Wname="W" # name of covariates

## specify split points for RRC method
Time_sep=c(0,15,25,35)

## specify the conditional values in the mediation effect measures
t.list=c(10,15,20,25,30,35,40,45); a0=0; a1=1; cw=0

## run the following code to perform the mediation analysis with approximate expressions
## Here, R=20 bootstrap replications were used for illustrative purposes
## In practice, one should try a large number (e.g., R=500)

## mediate_exact(example_data,Iname,Tname,Dname,Aname,
##                Asname,Mname,Wname,t.list,Time_sep,a0,a1,cw,R=20)
}
