% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_noerror.R
\name{mediate_unadj_approx}
\alias{mediate_unadj_approx}
\title{Mediation analysis based on the approximate expressions under a rare outcome assumption (no measurement error).}
\usage{
mediate_unadj_approx(data, Tname, Dname, Aname, Mname, Wname, a0, a1, cw)
}
\arguments{
\item{data}{dataset.}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}

\item{Aname}{name of the true exposure}

\item{Mname}{name of the mediator}

\item{Wname}{names of the covariates}

\item{a0}{baseline exposure level in the mediation effect measures}

\item{a1}{active exposure level in the mediation effect measures}

\item{cw}{levels of the covariates}
}
\value{
estimation of mediation effect measures based on the unadjusted method
}
\description{
Mediation analysis based on the approximate expressions under a rare outcome assumption (no measurement error).
}
