% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_noerror.R
\name{mediate_unadj_exact}
\alias{mediate_unadj_exact}
\title{Mediation analysis based on the exact expressions (no measurement error)}
\usage{
mediate_unadj_exact(
  data,
  Tname,
  Dname,
  Aname,
  Mname,
  Wname,
  t.list,
  a0,
  a1,
  cw,
  R = 50
)
}
\arguments{
\item{data}{dataset.}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}

\item{Aname}{name of the true exposure}

\item{Mname}{name of the mediator}

\item{Wname}{names of the covariates}

\item{t.list}{a list of time points for the mediation effect measures}

\item{a0}{baseline exposure level in the mediation effect measures}

\item{a1}{active exposure level in the mediation effect measures}

\item{cw}{levels of the covariates}

\item{R}{number of bootstrap numbers (default 50)}
}
\value{
estimation of mediation effect measures based on the unadjusted method
}
\description{
Mediation analysis based on the exact expressions (no measurement error)
}
