% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_noerror.R
\name{mediate_unadj_point_exact.f}
\alias{mediate_unadj_point_exact.f}
\title{Point estimation of the mediation effect based on exact expressions (no measurement error)}
\usage{
mediate_unadj_point_exact.f(
  Outcome.formula,
  Mediator.formula,
  data,
  Tname,
  Dname,
  Aname,
  Mname,
  Wname,
  a0,
  a1,
  cw,
  t.list
)
}
\description{
Point estimation of the mediation effect based on exact expressions (no measurement error)
}
