% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exact.R
\name{my.ghq}
\alias{my.ghq}
\title{Gauss-Hermite Quadrature Approximation (copied from spatstat package)}
\usage{
my.ghq(f, mu = 0, sd = 1, ..., order = 5)
}
\description{
Gauss-Hermite Quadrature Approximation (copied from spatstat package)
}
