% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exact.R
\name{orc2.ef}
\alias{orc2.ef}
\title{ORC2 method (for exact expressions)}
\usage{
orc2.ef(
  Outcome.formula,
  Mediator.formula,
  ME1.formula,
  ME2.formula,
  data,
  Iname,
  Tname,
  Dname,
  Aname,
  Asname,
  Mname,
  Wname,
  is.boot = 0
)
}
\arguments{
\item{Outcome.formula}{formula for the Cox outcome model.}

\item{Mediator.formula}{formula for the linear mediator model.}

\item{data}{dataset.}

\item{Iname}{name of the indicator on the main/validation study (I=1 for validation study sample, I=0 for main study samples).}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}
}
\value{
ORC2 estimates of the regression coefficients.
}
\description{
ORC2 method (for exact expressions)
}
