% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_noerror.R
\name{param_unadj_approx.f}
\alias{param_unadj_approx.f}
\title{Unadjusted method for regression coefficients (no measurement error)}
\usage{
param_unadj_approx.f(
  Outcome.formula,
  Mediator.formula,
  data,
  Tname,
  Dname,
  is.boot = 0
)
}
\arguments{
\item{Outcome.formula}{formula for the Cox outcome model.}

\item{Mediator.formula}{formula for the linear mediator model.}

\item{data}{dataset}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}
}
\value{
Unadjusted estimates of the regression coefficients.
}
\description{
Unadjusted method for regression coefficients (no measurement error)
}
