% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_approx.R
\name{unadj.f}
\alias{unadj.f}
\title{Unadjusted method that mistakenly treats the mismeasured exposure as the true exposure}
\usage{
unadj.f(
  Outcome.formula,
  Mediator.formula,
  data,
  Iname,
  Tname,
  Dname,
  is.boot = 0
)
}
\arguments{
\item{Outcome.formula}{formula for the Cox outcome model.}

\item{Mediator.formula}{formula for the linear mediator model.}

\item{data}{dataset.}

\item{Iname}{name of the indicator on the main/validation study (I=1 for validation study sample, I=0 for main study samples).}

\item{Tname}{name of the observed failure time.}

\item{Dname}{name of the failure indicator.}
}
\value{
Unadjusted estimates of the regression coefficients.
}
\description{
Unadjusted method that mistakenly treats the mismeasured exposure as the true exposure
}
